server/requirements.txt
grpcio>=1.10.0
grpcio-tools>=1.10.0
google-auth>=1.4.1
six>=1.11